Rm test_option_either.py: either now in main, move others into test_option.py
